implement coro scheduler
Implements a simple round-robin scheduler for coroutines run by a single thread.

Coroutines are powerful tools for multiplexing work in a single thread, but for C++ they can be a bit cumbersome. This is an attempt to wrap some of this complexity behind a simple API that runs all available coroutines until none are left or the scheduler is canceled.

Each coroutine runs until it is suspended. Then the scheduler checks whether or not the coroutine has returned. If so, then the scheduler removes it from its list of coroutines to run.

NOTE: Be aware that this is just an initial pass. The basic functionality is there but use it at your own risk.