Update directory structure for Python protos.
The current configuration led to naming conflicts when trying to use protos across project boundaries.
Python only imports a single module under a given name, which means that the directory structure that
the combination of Bazel's handling of external dependencies and the fact that `protoc` determines the
package name directly from the directory structure means that a proto residing at `rome/my.proto` will
not be found because there is already an `__init__.py` file in `external/rome`. To work around this, we
dump the protos into their own folder so that the imports do not conflict and the protos can actually
be found.